# trie, trie_prefetcher: don't copy when we are the exclusive owner of the trie